Clean up XML issues

+++ app/src/main/java/com/boardgamegeek/ui/widget/RangeSeekBar.java
@@ -440,7 +440,7 @@ public class RangeSeekBar<T extends Number> extends ImageView {
 	 */
 	private void drawThumb(float screenCoord, boolean pressed, Canvas canvas) {
 		canvas.drawBitmap(pressed ? thumbPressedImage : thumbImage, screenCoord - thumbHalfWidth,
-			(float) ((0.5f * getHeight()) - thumbHalfHeight), paint);
+			(0.5f * getHeight()) - thumbHalfHeight, paint);
 	}
 
 	/**
